Fix the pretrained resnet

+++ test/runtests.jl
@@ -99,7 +99,7 @@ end
     for (T, MODEL) in [
             (Float32, SqueezeNet),
             (Float32, VGG19),
-            (Float64, ResNet50),
+            (Float32, ResNet50),
             (Float64, DenseNet121),
             (Float64, GoogleNet)
         ]
